+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Changed
+- flow event listener for NC28
+
 ## 4.11.1 - 2023-11-20
 ### Fixed
 - save error for reports without table